backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
sentence-transformers==2.5.1
python-dotenv==1.0.1
pinecone-client==3.2.0
gunicorn==21.2.0